refactor: move github client under src/services/github/ and extract types
