+++ src/handlers/create-task.ts
@@ -1,6 +1,6 @@
 import type { CoderClient } from "../services/coder/client";
 import { TaskNameSchema } from "../services/coder/client";
-import type { GitHubClient } from "../github-client";
+import type { GitHubClient } from "../services/github/client";
 import type { Logger } from "../infra/logger";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
 import { generateTaskName } from "../task-utils";
