+++ src/handlers/pr-comment.ts
@@ -1,5 +1,5 @@
 import type { CoderClient } from "../services/coder/client";
-import type { GitHubClient } from "../github-client";
+import type { GitHubClient } from "../services/github/client";
 import type { Logger } from "../infra/logger";
 import { formatPRCommentMessage } from "../messages";
 import type { ActionOutputs, HandlerConfig } from "../config/handler-config";
